introduce transformation editor to pipelines refactor structure a bit remove unused TextAreaDataProvider remove seperate transformation route require successfully executed transformation for valid pipeline configuration
