+++ ui/src/App.vue
@@ -63,7 +63,6 @@ export default class App extends Vue {
     { title: 'Dashboard', route: '/' },
     { title: 'Datasources', route: '/datasources' },
     { title: 'Pipelines', route: '/pipelines' },
-    { title: 'Transformation', route: '/transformation' },
     { title: 'About', route: '/about' }
   ];
