+++ ui/src/main.ts
@@ -5,7 +5,7 @@ import router from './router'
 import Vuetify from 'vuetify'
 import 'vuetify/dist/vuetify.min.css'
 import '@mdi/font/css/materialdesignicons.css'
-import TransformationModule from '@/transformation/module'
+import TransformationModule from '@/pipeline/edit/transformation/module'
 import StorageModule from '@/storage/module'
 import AuthModule from '@/components/auth/module'
 import DatasourceModule from '@/datasource/datasource.module'
