+++ ui/src/router.ts
@@ -1,7 +1,6 @@
 import Vue from 'vue'
 import Router from 'vue-router'
 import Home from '@/views/Home.vue'
-import transformationRoutes from '@/transformation/router'
 import storageRoutes from '@/storage/router'
 import datasourceRoutes from '@/datasource/router'
 import pipelineRoutes from '@/pipeline/router'
@@ -28,7 +27,6 @@ let routes = [
   }
 ]
 
-routes = routes.concat(transformationRoutes)
 routes = routes.concat(pipelineRoutes)
 routes = routes.concat(datasourceRoutes)
 routes = routes.concat(storageRoutes)
