--- ui/src/transformation/interfaces/jobError.ts
@@ -1,7 +1,0 @@
-export default interface JobError {
-  name: string;
-  message: string;
-  lineNumber: number;
-  position: number;
-  stacktrace: string[];
-}
