--- ui/src/transformation/interfaces/jobResult.ts
@@ -1,13 +1,0 @@
-import JobError from './jobError'
-
-export default interface JobResult {
-  data?: object;
-  error?: JobError;
-  stats: Stats;
-}
-
-interface Stats {
-  durationInMilliSeconds: number;
-  startTimestamp: number;
-  endTimestamp: number;
-}
