--- ui/src/transformation/interfaces/transformationRequest.ts
@@ -1,4 +1,0 @@
-export default interface TransformationRequest {
-  func: string;
-  data: object;
-}
